requirements.txt
Flask==3.1.1
Flask-RESTful==0.3.10
Flask-Cors==6.0.1
mysql-connector-python==9.4.0
yfinance==0.2.65
blinker==1.9.0
click==8.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.4.0
Werkzeug==3.1.3